update migration: add quantity, update price to float

+++ migrations/sqls/20220103213823-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
 	id SERIAL PRIMARY KEY,
 	name VARCHAR NOT NULL,
-	price INT NOT NULL,
+	price FLOAT NOT NULL,
 	category VARCHAR
 );
